Translation

+++ src/main/java/com/example/demo4/Observer.java
@@ -1,6 +1,7 @@
 package com.example.demo4;
 
 public interface Observer {
-    void update(int x,int y,double zoom);
+    void update(Double x, Double y, double zoom);
+
     void update(String imgUrl);
 }
