+++ src/main/java/com/example/demo4/ThumbnailView.java
@@ -3,28 +3,28 @@ package com.example.demo4;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 
-public class ThumbnailView implements Observer{
+public class ThumbnailView implements Observer {
 
     private String imgUrl = "logo.png";
     private ImageView Iv;
 
-    public ThumbnailView(){
-        Image image = new Image(imgUrl,300,300,false,false);
+    public ThumbnailView() {
+        Image image = new Image(imgUrl, 300, 300, false, false);
         Iv = new ImageView(image);
     }
 
     @Override
-    public void update(int x, int y, double zoom) {
-        //ici fait rien avec cela car cette vue n'update pas la perspective
+    public void update(Double x, Double y, double zoom) {
+        // ici fait rien avec cela car cette vue n'update pas la perspective
     }
 
     @Override
     public void update(String imgUrl) {
         this.imgUrl = imgUrl; // change l'image
-        Iv.setImage(new Image(imgUrl,300,300,false,false));
+        Iv.setImage(new Image(imgUrl, 300, 300, false, false));
     }
 
-    public ImageView getView(){
+    public ImageView getView() {
         return Iv;
     }
 }
